telegram_celery_intergration/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
python-telegram-bot==20.8
gunicorn==21.2.0
whitenoise==6.6.0
PyJWT==2.10.1 